+++ R/mnis_extra.R
@@ -1,8 +1,8 @@
 
-#' mnis_Extra
+#' mnis_extra
 #'
 #'
-#' A wrapper for the \code{\link{mnis_Additional}} functions.
+#' A wrapper for the \code{\link{mnis_additional}} functions.
 #' Variable descriptions are taken from the mnis website: <http://data.parliament.uk/membersdataplatform/memberquery.aspx>
 #'
 #'
@@ -31,20 +31,17 @@
 #' @param Statuses Status history (e.g. suspensions and disqualifications) for a Member.
 #' @keywords mnis
 #' @examples \dontrun{
-#' x <- mnis_Extra(172)
+#' x <- mnis_extra(172)
 #'
 #' }
 #' @export
-#' @rdname mnis_Extra
-#' @seealso \code{\link{mnis_FullBiog}} \code{\link{mnis_BasicDetails}} \code{\link{mnis_Additional}}
+#' @rdname mnis_extra
+#' @seealso \code{\link{mnis_full_biog}} \code{\link{mnis_basic_details}} \code{\link{mnis_additional}}
 
-
-# http://data.parliament.uk/membersdataplatform/memberquery.aspx
-
-mnis_Extra <- function(ID, mem_id = TRUE, refDods = FALSE, Addresses = TRUE, BasicDetails = TRUE, BiographyEntries = TRUE, Committees = TRUE, 
-    Constituencies = TRUE, ElectionsContested = TRUE, Experiences = TRUE, GovernmentPosts = TRUE, Honours = TRUE, HouseMemberships = TRUE, 
-    Interests = TRUE, KnownAs = TRUE, MaidenSpeeches = TRUE, OppositionPosts = TRUE, OtherParliaments = TRUE, ParliamentaryPosts = TRUE, 
-    Parties = TRUE, PreferredNames = TRUE, Staff = TRUE, Statuses = TRUE) {
+mnis_extra <- function(ID, mem_id = TRUE, refDods = FALSE, Addresses = TRUE, BasicDetails = TRUE, BiographyEntries = TRUE, 
+    Committees = TRUE, Constituencies = TRUE, ElectionsContested = TRUE, Experiences = TRUE, GovernmentPosts = TRUE, 
+    Honours = TRUE, HouseMemberships = TRUE, Interests = TRUE, KnownAs = TRUE, MaidenSpeeches = TRUE, OppositionPosts = TRUE, 
+    OtherParliaments = TRUE, ParliamentaryPosts = TRUE, Parties = TRUE, PreferredNames = TRUE, Staff = TRUE, Statuses = TRUE) {
     
     ID <- as.character(ID)
     
@@ -53,75 +50,72 @@ mnis_Extra <- function(ID, mem_id = TRUE, refDods = FALSE, Addresses = TRUE, Bas
     } else {
         ID_Type <- "id="
     }
-    # Replace all this with calls to the specific functions, then combine into one data frame. Much easier, also gets around the
-    # call limit
+    # Replace all this with calls to the specific functions, then combine into one data frame. Much easier, also
+    # gets around the call limit
     mnis_DF_list <- list()
     
     if (Addresses == TRUE) 
-        Addresses_DF <- mnis_Addresses(ID)
+        Addresses_DF <- mnis_addresses(ID)
     mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
     if (BasicDetails == TRUE) 
-        BasicDetails_DF <- mnis_BasicDetails(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        BasicDetails_DF <- mnis_basic_details(ID)
+    mnis_DF_list[["BasicDetails_DF"]] <- Addresses_DF
     if (BiographyEntries == TRUE) 
         BiographyEntries_DF <- mnis_BiographyEntries(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+    mnis_DF_list[["BiographyEntries_DF"]] <- Addresses_DF
     if (Committees == TRUE) 
-        Committees_DF <- mnis_Committees(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Committees_DF <- mnis_committees(ID)
+    mnis_DF_list[["Committees_DF"]] <- Addresses_DF
     if (Constituencies == TRUE) 
-        Constituencies_DF <- mnis_Constituencies(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Constituencies_DF <- mnis_constituencies(ID)
+    mnis_DF_list[["Constituencies_DF"]] <- Addresses_DF
     if (ElectionsContested == TRUE) 
-        ElectionsContested_DF <- mnis_ElectionsContested(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        ElectionsContested_DF <- mnis_elections_contested(ID)
+    mnis_DF_list[["ElectionsContested_DF"]] <- Addresses_DF
     if (Experiences == TRUE) 
-        Experiences_DF <- mnis_Experiences(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Experiences_DF <- mnis_experiences(ID)
+    mnis_DF_list[["Experiences_DF"]] <- Addresses_DF
     if (GovernmentPosts == TRUE) 
-        GovernmentPosts_DF <- mnis_GovernmentPosts(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        GovernmentPosts_DF <- mnis_government_posts(ID)
+    mnis_DF_list[["GovernmentPosts_DF"]] <- Addresses_DF
     if (Honours == TRUE) 
-        Honours_DF <- mnis_Honours(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Honours_DF <- mnis_honours(ID)
+    mnis_DF_list[["Honours_DF"]] <- Addresses_DF
     if (HouseMemberships == TRUE) 
-        HouseMemberships_DF <- mnis_HouseMemberships(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        HouseMemberships_DF <- mnis_house_memberships(ID)
+    mnis_DF_list[["HouseMemberships_DF"]] <- Addresses_DF
     if (Interests == TRUE) 
-        Interests_DF <- mnis_Interests(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Interests_DF <- mnis_interests(ID)
+    mnis_DF_list[["Interests_DF"]] <- Addresses_DF
     if (KnownAs == TRUE) 
-        KnownAs_DF <- mnis_KnownAs(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        KnownAs_DF <- mnis_known_as(ID)
+    mnis_DF_list[["KnownAs_DF"]] <- Addresses_DF
     if (MaidenSpeeches == TRUE) 
-        MaidenSpeeches_DF <- mnis_MaidenSpeeches(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        MaidenSpeeches_DF <- mnis_maiden_speeches(ID)
+    mnis_DF_list[["MaidenSpeeches_DF"]] <- Addresses_DF
     if (OppositionPosts == TRUE) 
-        OppositionPosts_DF <- mnis_OppositionPosts(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        OppositionPosts_DF <- mnis_opposition_posts(ID)
+    mnis_DF_list[["OppositionPosts_DF"]] <- Addresses_DF
     if (OtherParliaments == TRUE) 
-        OtherParliaments_DF <- mnis_OtherParliaments(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        OtherParliaments_DF <- mnis_other_parliaments(ID)
+    mnis_DF_list[["OtherParliaments_DF"]] <- Addresses_DF
     if (ParliamentaryPosts == TRUE) 
-        ParliamentaryPosts_DF <- mnis_ParliamentaryPosts(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        ParliamentaryPosts_DF <- mnis_parliamentary_posts(ID)
+    mnis_DF_list[["ParliamentaryPosts_DF"]] <- Addresses_DF
     if (Parties == TRUE) 
-        Parties_DF <- mnis_Parties(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        Parties_DF <- mnis_parties(ID)
+    mnis_DF_list[["Parties_DF"]] <- Addresses_DF
     if (PreferredNames == TRUE) 
-        PreferredNames_DF <- mnis_PreferredNames(ID)
-    mnis_DF_list[["Addresses_DF"]] <- Addresses_DF
+        PreferredNames_DF <- mnis_preferred_names(ID)
+    mnis_DF_list[["PreferredNames_DF"]] <- Addresses_DF
     if (Staff == TRUE) 
-        Staff_DF <- mnis_Staff(ID)
+        Staff_DF <- mnis_staff(ID)
     mnis_DF_list[["Staff_DF"]] <- Staff_DF
     if (Statuses == TRUE) 
-        Statuses_DF <- mnis_Statuses(ID)
+        Statuses_DF <- mnis_statuses(ID)
     mnis_DF_list[["Statuses_DF"]] <- Statuses_DF
-    
     
     x <- plyr::join_all(mnis_DF_list, match = "first")
     
     x
 }
-
-