+++ R/mnis_member_date.R
@@ -1,4 +1,4 @@
-#' mnis_MemberDate
+#' mnis_member_date
 #'
 #' Returns a data frame with a members status on a given date.
 #' @param ID The ID of the member.
@@ -7,11 +7,11 @@
 #' @keywords mnis
 #' @export
 #' @examples \dontrun{
-#' x <- mnis_MemberDate(172)
+#' x <- mnis_member_date(172)
 #'
 #' }
 
-mnis_MemberDate <- function(ID = NULL, Date = Sys.Date()) {
+mnis_member_date <- function(ID = NULL, Date = Sys.Date()) {
 
     ID <- as.character(ID)
 
@@ -29,4 +29,12 @@ mnis_MemberDate <- function(ID = NULL, Date = Sys.Date()) {
 
     x <- as.list(got$Member)
 
+}
+
+#' @export
+#' @rdname mnis_member_date
+#' @usage NULL
+mnis_MemberDate <- function(ID = NULL, Date = Sys.Date()) {
+    .Deprecated("mnis_member_date")
+    mnis_member_date(ID = ID, Date = Date)
 }